requirements.txt
Faker~=37.5.3
selenium~=4.14.0
pytest~=8.4.1
allure-pytest~=2.15.0